+++ coldfront/config/settings.py
@@ -1,197 +1,53 @@
-"""
-Default Django settings for ColdFront project.
-"""
 import os
-import sys
+from split_settings.tools import optional, include
+from coldfront.config.env import ENV, PROJECT_ROOT
 
-from django.contrib.messages import constants as messages
-
-#------------------------------------------------------------------------------
-# Django config for ColdFront
-#------------------------------------------------------------------------------
-
-BASE_DIR = os.path.dirname(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
-ALLOWED_HOSTS = ['*']
-
-#------------------------------------------------------------------------------
-# Django Apps
-#------------------------------------------------------------------------------
-INSTALLED_APPS = [
-    'django.contrib.admin',
-    'django.contrib.auth',
-    'django.contrib.contenttypes',
-    'django.contrib.sessions',
-    'django.contrib.messages',
-    'django.contrib.staticfiles',
-    'django.contrib.humanize',
+# ColdFront split settings
+coldfront_configs = [
+    'base.py',
+    'database.py',
+    'auth.py',
+    'logging.py',
+    'core.py',
 ]
 
+if ENV.bool('EMAIL_ENABLED', default=False):
+    coldfront_configs.append('email.py')
 
-# Additional Apps
-INSTALLED_APPS += [
-    'crispy_forms',
-    'sslserver',
-    'django_q',
-    'simple_history',
-]
-
-# ColdFront Apps
-INSTALLED_APPS += [
-    'coldfront.core.user',
-    'coldfront.core.field_of_science',
-    'coldfront.core.utils',
-    'coldfront.core.portal',
-    'coldfront.core.project',
-    'coldfront.core.resource',
-    'coldfront.core.allocation',
-    'coldfront.core.grant',
-    'coldfront.core.publication',
-    'coldfront.core.research_output',
-]
-
-#------------------------------------------------------------------------------
-# Django Middleware
-#------------------------------------------------------------------------------
-MIDDLEWARE = [
-    'django.middleware.security.SecurityMiddleware',
-    'django.contrib.sessions.middleware.SessionMiddleware',
-    'django.middleware.common.CommonMiddleware',
-    'django.middleware.csrf.CsrfViewMiddleware',
-    'django.contrib.auth.middleware.AuthenticationMiddleware',
-    'django.contrib.messages.middleware.MessageMiddleware',
-    'django.middleware.clickjacking.XFrameOptionsMiddleware',
-    'simple_history.middleware.HistoryRequestMiddleware',
-]
-
-#------------------------------------------------------------------------------
-# Database settings
-#------------------------------------------------------------------------------
-DATABASES = {
-    'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': os.path.join(BASE_DIR, 'coldfront.db'),
-    }
+# ColdFront plugin settings
+plugin_configs = {
+    'PLUGIN_SLURM': 'plugins/slurm.py',
+    'PLUGIN_IQUOTA': 'plugins/iquota.py',
+    'PLUGIN_FREEIPA': 'plugins/freeipa.py',
+    'PLUGIN_SYSMON': 'plugins/system_montior.py',
+    'PLUGIN_XDMOD': 'plugins/xdmod.py',
+    'PLUGIN_AUTH_OIDC': 'plugins/openid.py',
+    'PLUGIN_AUTH_LDAP': 'plugins/ldap.py',
 }
 
-#------------------------------------------------------------------------------
-# Authentication backends
-#------------------------------------------------------------------------------
-AUTHENTICATION_BACKENDS = [
-    'django.contrib.auth.backends.ModelBackend',
+# This allows plugins to be enabled via environment variables. Can alternatively
+# add the relevant configs to local_settings.py
+for key, pc in plugin_configs.items():
+    if ENV.bool(key, default=False):
+        coldfront_configs.append(pc)
+
+# Local settings overrides
+local_configs = [
+    # Local settings relative to coldfront.config package
+    'local_settings.py',
+
+     # System wide settings for production deployments
+    '/etc/coldfront/local_settings.py',
+
+    # Local settings relative to coldfront project root 
+    PROJECT_ROOT('local_settings.py')
 ]
 
-#------------------------------------------------------------------------------
-# Django site settings
-#------------------------------------------------------------------------------
-ROOT_URLCONF = 'coldfront.config.urls'
+if ENV.str('COLDFRONT_CONFIG', default='') != '':
+    # Local settings from path specified via environment variable
+    local_configs.append(environ.Path(ENV.str('COLDFRONT_CONFIG'))())
 
-TEMPLATES = [
-    {
-        'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [
-            os.path.join(BASE_DIR, 'site/templates'),
-            '/usr/share/coldfront/site/templates',
-            os.path.join(BASE_DIR, 'coldfront/templates'),
-        ],
-        'APP_DIRS': True,
-        'OPTIONS': {
-            'context_processors': [
-                'django.template.context_processors.debug',
-                'django.template.context_processors.request',
-                'django.contrib.auth.context_processors.auth',
-                'django.contrib.messages.context_processors.messages',
-                'django_settings_export.settings_export',
-            ],
-        },
-    },
-]
+for lc in local_configs:
+    coldfront_configs.append(optional(lc))
 
-SESSION_COOKIE_AGE = 60 * 15
-SESSION_SAVE_EVERY_REQUEST = True
-SESSION_COOKIE_SAMESITE  = 'Strict'
-
-WSGI_APPLICATION = 'coldfront.config.wsgi.application'
-
-USE_I18N = True
-USE_L10N = True
-USE_TZ = True
-
-ADMIN_COMMENTS_SHOW_EMPTY = True
-
-MESSAGE_TAGS = {
-    messages.DEBUG: 'info',
-    messages.INFO: 'info',
-    messages.SUCCESS: 'success',
-    messages.WARNING: 'warning',
-    messages.ERROR: 'danger',
-}
-
-CRISPY_TEMPLATE_PACK = 'bootstrap4'
-
-STATIC_URL = '/static/'
-STATIC_ROOT = os.path.join(BASE_DIR, 'static_root')
-STATICFILES_DIRS = [
-    os.path.join(BASE_DIR, 'coldfront/static'),
-]
-
-# Add local site static files
-if os.path.isdir(os.path.join(BASE_DIR, 'site/static')):
-    STATICFILES_DIRS.insert(0, os.path.join(BASE_DIR, 'site/static'))
-
-# Add system site static files
-if os.path.isdir('/usr/share/coldfront/site/static'):
-    STATICFILES_DIRS.insert(0, '/usr/share/coldfront/site/static')
-
-LOGIN_URL = '/user/login'
-LOGIN_REDIRECT_URL = '/'
-LOGOUT_REDIRECT_URL = '/'
-
-SU_LOGIN_CALLBACK = "coldfront.core.utils.common.su_login_callback"
-SU_LOGOUT_REDIRECT_URL = "/admin/auth/user/"
-
-
-SETTINGS_EXPORT = []
-
-#------------------------------------------------------------------------------
-# Local settings overrides (see local_settings.py.sample)
-#------------------------------------------------------------------------------
-try:
-    from coldfront.config.local_strings import *
-except ImportError:
-    print("local_strings.py file is required. Copy coldfront/config/local_strings.py.sample to local_strings.py")
-    sys.exit()
-
-try:
-    from coldfront.config.local_settings import *
-except ImportError:
-    print("local_settings.py file is required. Copy coldfront/config/local_settings.py.sample to local_settings.py")
-    sys.exit()
-
-try:
-    INSTALLED_APPS = INSTALLED_APPS + EXTRA_APPS
-except NameError:
-    INSTALLED_APPS = INSTALLED_APPS
-
-try:
-    MIDDLEWARE = MIDDLEWARE + EXTRA_MIDDLEWARE
-except NameError:
-    MIDDLEWARE = MIDDLEWARE
-
-try:
-    AUTHENTICATION_BACKENDS = AUTHENTICATION_BACKENDS + EXTRA_AUTHENTICATION_BACKENDS
-except NameError:
-    AUTHENTICATION_BACKENDS = AUTHENTICATION_BACKENDS
-
-
-if 'django_su.backends.SuBackend' in EXTRA_AUTHENTICATION_BACKENDS:
-    INSTALLED_APPS.insert(0, 'django_su')
-    TEMPLATES[0]['OPTIONS']['context_processors'].extend(['django_su.context_processors.is_su', ])
-
-import coldfront
-VERSION = coldfront.__version__
-
-try:
-    SETTINGS_EXPORT = SETTINGS_EXPORT + LOCAL_SETTINGS_EXPORT
-except NameError:
-    SETTINGS_EXPORT = SETTINGS_EXPORT 
+include(*coldfront_configs)