+++ coldfront/core/portal/views.py
@@ -41,17 +41,15 @@ def home(request):
         context['project_list'] = project_list
         context['allocation_list'] = allocation_list
         try:
-            if(settings.ONDEMAND_URL != None):
-                print(settings.ONDEMAND_URL)
-                context['ondemand_url'] = settings.ONDEMAND_URL
+            context['ondemand_url'] = settings.ONDEMAND_URL
         except AttributeError:
-            print("No Ondemand URL provided")
+            pass
     else:
         template_name = 'portal/nonauthorized_home.html'
 
-    context['EXTRA_APPS'] = settings.EXTRA_APPS
+    context['EXTRA_APPS'] = settings.INSTALLED_APPS
 
-    if 'coldfront.plugins.system_monitor' in settings.EXTRA_APPS:
+    if 'coldfront.plugins.system_monitor' in settings.INSTALLED_APPS:
         from coldfront.plugins.system_monitor.utils import get_system_monitor_context
         context.update(get_system_monitor_context())