+++ coldfront/core/publication/forms.py
@@ -6,7 +6,7 @@ from coldfront.core.publication.models import PublicationSource
 class PublicationAddForm(forms.Form):
     title = forms.CharField(max_length=1024, required=True)
     author = forms.CharField(max_length=1024, required=True)
-    year = forms.CharField(max_length=4, required=True)
+    year = forms.IntegerField(min_value=1500, max_value=2090, required=True)
     journal = forms.CharField(max_length=1024, required=True)
     source = forms.CharField(widget=forms.HiddenInput())  # initialized by view