+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## [Unreleased]
 
+- Fix [#235](https://github.com/ubccr/coldfront/issues/235)
+
+## [1.0.3] - 2021-03-02
+
+- Refactor ColdFront settings. See [PR #264](https://github.com/ubccr/coldfront/pull/264)
+- Lots of documenation updates. [See here](https://coldfront.readthedocs.io)
+- Fix setuptools for pip installs
+
 ## [1.0.2] - 2021-02-15
 
 - Accessibility fixes @zooley
@@ -35,7 +43,8 @@
 
 - Initial release
 
-[Unreleased]: https://github.com/ubccr/coldfront/compare/v1.0.1...HEAD
 [0.0.1]: https://github.com/ubccr/coldfront/releases/tag/v0.0.1
 [1.0.1]: https://github.com/ubccr/coldfront/releases/tag/v1.0.1
 [1.0.2]: https://github.com/ubccr/coldfront/releases/tag/v1.0.2
+[1.0.3]: https://github.com/ubccr/coldfront/releases/tag/v1.0.3
+[Unreleased]: https://github.com/ubccr/coldfront/compare/v1.0.3...HEAD